project-plan-generator/pyproject.toml
[tool.poetry]
name = "project-plan-generator"
version = "0.1.0"
description = "LangGraph-based agentic AI system for generating comprehensive project plans from requirements"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-anthropic = "^0.2.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
typing-extensions = "^4.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
ruff = "^0.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
